playlist service work

+++ docker/node/modules/playlist/service.js
@@ -1,58 +1,87 @@
 const { actions } = require("../auth");
-const { getSocketName } = require("../sessions");
+const { getSocketName, userTypes } = require("../sessions");
 const { ServiceBase } = require("../base");
 const { events } = require("../log/events");
 const { Playlist } = require("./playlist");
 const { Video, VideoFormat } = require("./video");
 
 const settings = require("../../bt_data/settings");
-const config = require("../../bt_data/db_info");
 
+const { YoutubeHandler } = require("./handlers/youtube");
+const { VimeoHandler } = require("./handlers/vimeo");
+const { DailymotionHandler } = require("./handlers/dailymotion");
+const { SoundcloudHandler } = require("./handlers/soundcloud");
+const { RedditHandler } = require("./handlers/reddit");
+const { DashHandler } = require("./handlers/dash");
+const { HLSHandler } = require("./handlers/hls");
+const { FileHandler } = require("./handlers/file");
+const { ManifestHandler } = require("./handlers/manifest");
+const { TwitchHandler } = require("./handlers/twitch");
+const { TwitchClipHandler } = require("./handlers/twitchclip");
+
+const PlayerState = {
+	RUNNING: 1,
+	PAUSED: 2
+};
 
 exports.PlaylistService = class extends ServiceBase {
-	get current() {
-		return this.playlist.current();
-	}
-
-	get cursor() {
-		return this.playlist.cursor();
-	}
-
-	constructor(services) {
+	constructor(services, eventServer) {
 		super(services);
+
 		this.auth = services.auth;
 		this.io = services.io;
 		this.db = services.db;
-		this.sessions = services.sessions;
 		this.playlist = new Playlist();
 		this.log = services.log;
+		this.eventServer = eventServer;
+		this.sessions = services.sessions;
 
 		this.exposeSocketActions({
 			playNext: this.advance.bind(this),
 			sortPlaylist: this.move.bind(this),
 			forceVideoChange: this.jump.bind(this),
 			addVideo: this.add.bind(this),
+			delVideo: this.remove.bind(this),
 			fondleVideo: this.fondle.bind(this),
+			myPlaylistIsInited: this.onPlaylistReady.bind(this),
+			forceStateChange: this.pause.bind(this),
+			videoSeek: this.seek.bind(this),
+			removeLeader: this.resumeOwnership.bind(this)
 		});
 
 		this.time = 0;
-		//FIXME: Change to a "enum"
-		this.controller = 'server';
-		this.grabbers = new Map();
-		this.handlers = new Map(
-			[VideoFormat.YOUTUBE, null]
-		);
+		this.state = PlayerState.RUNNING;
+
+		//otherwise same but playlist is the actual one, not this service
+		//probably not needed, but meh
+		const links = {
+			...services,
+			playlist: this.playlist
+		};
+
+		this.handlers = new Map([
+			[VideoFormat.YOUTUBE, new YoutubeHandler(links)],
+			[VideoFormat.VIMEO, new VimeoHandler(links)],
+			[VideoFormat.DAILYMOTION, new DailymotionHandler(links)],
+			[VideoFormat.SOUNDCLOUD, new SoundcloudHandler(links)],
+			[VideoFormat.REDDIT, new RedditHandler(links)],
+			[VideoFormat.DASH, new DashHandler(links)],
+			[VideoFormat.HLS, new HLSHandler(links)],
+			[VideoFormat.FILE, new FileHandler(links)],
+			[VideoFormat.MANIFEST, new ManifestHandler(links)],
+			[VideoFormat.TWITCH, new TwitchHandler(links)],
+			[VideoFormat.TWITCHCLIP, new TwitchClipHandler(links)],
+		]);
+
 		this.paused = false;
+		this.current = null;
 	}
 
 	async init() {
-		//TODO: Remove these once parity has been achieved
-		//and no shenanigans happen during normal use
-		await this.db.query`DROP TABLE IF EXISTS videos_new`;
-		await this.db.query`CREATE TABLE videos_new LIKE videos`;
+		super.init();
 
 		const { result: videos } = await this.db.query`
-			select * from videos_new order by position
+			select * from videos order by position
 		`;
 
 		const {result: [active, time]} = await this.db.query`
@@ -64,15 +93,24 @@ exports.PlaylistService = class extends ServiceBase {
 
 		this.playlist.initialise(
 			videos.map(row => new Video(row)),
-			active,
+			active.value,
 		);
+
+		//wasteful but meh, ensures we always reset positions if something
+		//went wrong, also only done once when server starts
+		for (const [index, video] of this.playlist.videos().entries()) {
+			await this.db.query`
+				update videos set position = ${index} where videoid = ${video.id()}
+			`;
+		}
 		
-		this.time = Number.parseInt(time || -settings.vc.head_time, 10);
+		this.time = Number.parseInt(time.value || -settings.vc.head_time, 10);
+		this.current = this.playlist.current();
 	}
 
 	async advance(socket) {
 		this.current.removeTag(true);
-		
+
 		if (this.current.volatile()) {
 			await this.remove(socket, {pos: this.cursor, sanityid: this.current.id()});
 		} else {
@@ -80,76 +118,38 @@ exports.PlaylistService = class extends ServiceBase {
 		}
 
 		//set new video and reset time
+		this.current = this.playlist.current();
 		this.time = -settings.vc.head_time;
 		this.log.info(events.EVENT_VIDEO_CHANGE,
 			"changed video to {videoTitle}",
 			{ videoTitle: this.current.title() });
 
-		//send videochange to eventserver
-		this.events.emit('videoChange', {
-			id: this.current.id(),
-			length: this.current.duration(),
-			title: this.current.title(),
-			type: this.current.source(),
-			volat: this.current.volatile()
-		});
+		this.announce(this.io.sockets, "forceVideoChange");
 	}
 
 	async add(socket, data) {
-		if (!this.auth.can(socket, actions.ACTION_CONTROL_PLAYLIST)) {
+		if (!this.auth.can(socket.session, actions.ACTION_CONTROL_PLAYLIST)) {
 			throw new Error("User has no permission to add videos");
 		}
 
-		if (!this.grabbers.has(data.videotype)) {
+		if (!this.handlers.has(data.videotype)) {
 			throw new Error("Format or provider has no handler implemented");
 		}
-		
-		const handler = this.handlers.get(data.videotype);
-		const video = await handler.handle(data);
 
-		const position = data.queue ? this.cursor + 1 : this.playlist.videos().length;
+		const video = await this.handlers.get(data.videotype).handle(
+			socket,
+			data
+		).catch((err) => socket.emit('dupeAdd', err));
 
-		if (data.queue) {
-			this.playlist.insert(video, position);
-		} else {
-			this.playlist.append(video);
-		}
-
-		//remove it from the videos_history (why?)
-		await this.db.query`
-			DELETE FROM
-				videos_history
-			WHERE
-				videoid = ${video.id()}
-		`;
-
-		//insert the actual video to table
-		await this.db.query`
-			insert into videos_new 
-				(position, videoid, videotitle, videolength, videotype, videovia, meta) 
-			VALUES 
-				(
-					${position}, 
-					${video.id()}, 
-					${video.title()}, 
-					${video.duration()}, 
-					${video.source()},
-					${getSocketName(socket)}
-					${JSON.stringify(video.metadata())}
-				)
-		`;
-
-		//update positions only if we queued as next video
-		if (data.queue) {
-			await this.db.query`
-				UPDATE videos_new
-				SET position = position + 1
-				WHERE position >= ${position} AND id IS NOT ${video.id()}
-			`;
-		}
+		this.io.sockets.emit('addVideo', {
+			queue: data.queue, 
+			video: video.pack(), 
+			sanityid: this.current.id()
+		});
 	}
 
 	resync(socket) {
+		socket.emit('doorStuck');
 		socket.emit('refreshPlaylist', {desynced: true, items: this.items});
 	}
 
@@ -162,15 +162,16 @@ exports.PlaylistService = class extends ServiceBase {
 
 		const min = Math.min(data.from, data.to);
 		const max = Math.max(data.from, data.to);
+		const sign = -Math.sign(data.to - data.from);
 
 		await this.db.query`
-			UPDATE videos_new SET position = ${data.to} WHERE position = ${data.from};
+			UPDATE videos SET position = ${data.to} WHERE position = ${data.from};
 		`;
 
 		await this.db.query`
-			UPDATE videos_new 
-			SET position = position + ${Math.sign(data.to - data.from)}
-			WHERE position >= ${min} and position <= ${max} and id is not ${data.sanityid}
+			UPDATE videos
+			SET position = position + ${sign}
+			WHERE position >= ${min} and position <= ${max} and videoid <> ${data.sanityid}
 		`;
 
 		this.io.sockets.emit("sortPlaylist", data);
@@ -184,29 +185,39 @@ exports.PlaylistService = class extends ServiceBase {
 		const params = [
 			video.id(),
 			video.title(),
-			video.length(),
+			video.duration(),
 			video.source(),
-			"NOW()",
-			JSON.stringify(video.meta()),
+			JSON.stringify(video.metadata()),
 		];
 
-		await this.db.query`
-			insert into videos_history (videoid, videotitle, videolength, videotype, date_added, meta) 
-			values (${params.join(",")})
-		`;
+		await this.db.query(
+			[`insert into videos_history (videoid, videotitle, videolength, videotype, date_added, meta) values (?,?,?,?,NOW(),?)`],
+			...params
+		);
 	}
 
 	async remove(socket, data) {
-		const video = this.playlist.at(data.pos);
+		if (!this.auth.can(socket.session, actions.ACTION_DELETE_VIDEO)) {
+			return;
+		}
+
+		const active = this.playlist.getCursor() === data.index;
+		const video = this.playlist.at(data.index);
 
 		if (video.id() !== data.sanityid) {
 			return this.resync(socket);
 		}
 
-		this.playlist.remove(data.pos);
+		//also handles position change
+		this.playlist.remove(data.index);
 
+		if (active) {
+			this.current = this.playlist.current();
+			this.announce(this.io.sockets, 'forceVideoChange');
+		}
+		
 		await this.db.query`
-			delete from videos_new where videoid = ${video.id()} limit 1
+			delete from videos where videoid = ${video.id()} limit 1
 		`
 			.then(() => this.store(video))
 			.catch((e) => this.log.error(
@@ -216,42 +227,50 @@ exports.PlaylistService = class extends ServiceBase {
 				e,
 			));
 
-		//shift positions
+		//update positions
 		await this.db.query`
-			update videos_new set position = position - 1 where position > ${data.pos} 
+			update videos set position = position - 1 where position > ${data.pos} 
 		`;
 
 		this.io.sockets.emit("delVideo", data);
-		this.log.info(events.EVENT_ADMIN_MOVED_VIDEO, "NEW: Cursor is at: {cursor}", { cursor: this.cursor });
+		this.log.info(events.EVENT_ADMIN_MOVED_VIDEO, "NEW: Cursor is at: {cursor}", { cursor: this.playlist.getCursor() });
 	}
 
 	async jump(socket, data) {
+		if (!this.auth.can(socket.session, actions.ACTION_CONTROL_PLAYLIST)) {
+			throw new Error("User has no permission to control playlist");
+		}
+
 		if (this.playlist.at(data.index).id() !== data.sanityid) {
 			return this.resync(socket);
 		}
-
+		
+		//remove volatile tag
 		this.current.removeTag(true);
 
-		if (this.current.volatile()) {
-			await this.delete(socket, {pos: this.cursor, sanityid: data.sanityid});
-			data.index -= 1;
-		}
-
-		this.playlist.set_cursor(data.index);
+		//we need these two things to handle deletion after jumping
+		//so that we don't need to have special handling
+		//for deleting active video (which is already handled)
+		const old = this.playlist.current();
+		const cursor = this.playlist.getCursor();
+		
+		this.playlist.setCursor(data.index);
+		this.time = -settings.vc.head_time;
+		this.current = this.playlist.current();
 
 		this.announce(this.io.sockets, "forceVideoChange");
-		this.log.info(events.EVENT_ADMIN_FORCED_VIDEO_CHANGE, "NEW: {mod} forced video change", {
-			mod: getSocketName(socket),
-			type: "playlist",
-		});
+
+		if (old.volatile()) {
+			await this.remove(socket, {index: cursor, sanityid: old.sanityid});
+		}
 	}
 
 	async fondle(socket, data) {
 		if (!this.auth.can(socket.session, actions.ACTION_SET_VIDEO_VOLATILE)) {
-			return;
+			throw new Error("User has no permission to fondle videos");
 		}
 
-		if (!this.current.id() !== data.sanityid) {
+		if (this.current.id() !== data.sanityid) {
 			return this.resync(socket);
 		}
 
@@ -273,24 +292,75 @@ exports.PlaylistService = class extends ServiceBase {
 
 	announce(socket, event) {
 		socket.emit(event, {
-			video: this.current.pack(),
+			video: this.playlist.current().pack(),
 			time: this.time,
 			state: this.state,
-		});
-
-		this.eventServer.emit('videoStatus', {
-			time: Math.round(this.time),
-			state: this.state
 		});
 	}
 
 	seek(socket, event) {
 		if (!this.auth.can(socket.session, actions.ACTION_CONTROL_VIDEO)) {
+			throw new Error("User has no permission to control video");
+		}
+
+		this.time = Math.ceil(event);
+		this.log.info(events.EVENT_ADMIN_FORCED_VIDEO_CHANGE, "{mod} seeked to {event}", {
+			mod: getSocketName(socket),
+			event
+		});
+
+		this.announce(this.io.sockets, 'hbVideoDetail');
+		this.db.upsertMisc('server_time', this.time);
+	}
+
+	pause(socket, data) {
+		if (!this.auth.can(socket.session, actions.ACTION_CONTROL_VIDEO)) {
+			throw new Error("User has no permission to control video");
+		}
+
+		this.log.info(events.EVENT_ADMIN_FORCED_VIDEO_CHANGE, "{mod} paused video at {event} with videolength of {duration}", {
+			mod: getSocketName(socket),
+			event: this.time,
+			duration: this.current.duration()
+		});
+		
+		this.paused = data.state === PlayerState.PAUSED;
+		this.announce(this.io.sockets, 'forceStateChange');
+	}
+
+	//resume ownership if the video was paused and last (modded) berry is gone
+	resumeOwnership() {
+		if (!this.paused) {
 			return;
 		}
 
-		this.time = event;
-		this.announce(this.io.sockets, 'hbVideoDetail');
+		const berries = this.sessions.getBerries().filter(u => u.type >= userTypes.MODERATOR);
+
+		if (berries.length === 0) {
+			this.paused = false;
+		}
+	}
+
+	shuffle(socket) {
+		if (!this.auth.can(socket.session, actions.ACTION_RANDOMIZE_LIST)) {
+			throw new Error("User has no permission to randomize playlist");
+		}
+
+		this.playlist.shuffle();
+		this.io.sockets.emit('recvNewPlaylist', this.playlist.pack());
+	}
+
+	onPlaylistReady(socket) {
+		this.announce(socket.session, 'createPlayer');
+	}
+
+	onSocketConnected(socket) {
+		super.onSocketConnected(socket);
+
+		socket.session.emit(
+			"recvPlaylist",
+			this.playlist.pack()
+		);
 	}
 
 	onTick(elapsed) {
@@ -298,23 +368,18 @@ exports.PlaylistService = class extends ServiceBase {
 			return;
 		}
 
-		this.time += Math.round(elapsed / 1000);
+		//always assume we advance 1s
+		this.time += Math.ceil(elapsed / 1000);
 
 		if (this.time + 1 >= this.current.duration() + settings.vc.tail_time) {
 			this.advance();
+			this.db.upsertMisc('server_active_videoid', this.current.id());
 			return;
 		}
 
-		if (this.time > 0 && this.time % (settings.core.heartbeat_interval / 1000)) {
+		if (this.time > 0 && this.time % (settings.core.heartbeat_interval / 1000) === 0) {
 			this.announce(this.io.sockets, "hbVideoDetail");
-			this.db.query`
-				UPDATE
-					misc
-				SET
-					value = ${this.time}
-				WHERE
-					name = 'server_time'
-			`;
+			this.db.upsertMisc('server_time', this.time);
 		}
 	}
 };
